chore: re-org test files
