+++ test/parser/errors.test.ts
@@ -1,28 +1,6 @@
-import { schema } from '../src/parser/schema';
-import { Template } from '../src/parser/template';
-import { Testing } from './util';
-
-test('invalid enum option raises an error', async () => {
-  // GIVEN
-  const template = {
-    Resources: {
-      Hello: {
-        Type: 'aws-cdk-lib.aws_sqs.Queue',
-        Properties: {
-          encryption: 'boom',
-        },
-      },
-    },
-  };
-
-  // THEN
-  expect(template).not.toBeValidTemplate();
-  await expect(
-    Testing.synth(await Template.fromObject(template), false)
-  ).rejects.toThrow(
-    'Expected choice for enum type aws-cdk-lib.aws_sqs.QueueEncryption to be one of UNENCRYPTED|KMS_MANAGED|KMS, got: boom'
-  );
-});
+import { schema } from '../../src/parser/schema';
+import { Template } from '../../src/parser/template';
+import '../util';
 
 test('invalid tags raises an error', async () => {
   // GIVEN
