+++ test/type-resolution/callables.test.ts
@@ -1,9 +1,7 @@
 import * as reflect from 'jsii-reflect';
 import { Template } from '../../src/parser/template';
-import { StructExpression } from '../../src/type-resolution/struct';
 import { TypedTemplate } from '../../src/type-resolution/template';
-import { getCdkConstruct } from '../template';
-import { Testing } from '../util';
+import { matchConstruct, Testing } from '../util';
 
 let typeSystem: reflect.TypeSystem;
 
@@ -33,16 +31,15 @@ test('Static Methods are resolved correctly', async () => {
   const typedTemplate = new TypedTemplate(template, { typeSystem });
 
   // THEN
-  const myLambda = getCdkConstruct(typedTemplate, 'MyLambda');
-  expect(myLambda.type).toBe('construct');
-  expect(myLambda.props?.type).toBe('struct');
-  expect((myLambda.props as StructExpression)?.fields).toHaveProperty(
-    'code',
-    expect.objectContaining({
-      type: 'staticMethodCall',
-      fqn: 'aws-cdk-lib.aws_lambda.Code',
-      namespace: 'aws_lambda',
-      method: 'fromAsset',
+  const myLambda = typedTemplate.resource('MyLambda');
+  expect(myLambda).toEqual(
+    matchConstruct({
+      code: expect.objectContaining({
+        type: 'staticMethodCall',
+        fqn: 'aws-cdk-lib.aws_lambda.Code',
+        namespace: 'aws_lambda',
+        method: 'fromAsset',
+      }),
     })
   );
   expect(template.template).toBeValidTemplate();
@@ -80,14 +77,15 @@ test('Can provide implementation via static method', async () => {
   const typedTemplate = new TypedTemplate(template, { typeSystem });
 
   // THEN
-  const myLambda = getCdkConstruct(typedTemplate, 'MyFleet');
-  expect((myLambda.props as StructExpression)?.fields).toHaveProperty(
-    'machineImage',
-    expect.objectContaining({
-      type: 'staticMethodCall',
-      fqn: 'aws-cdk-lib.aws_ecs.EcsOptimizedImage',
-      namespace: 'aws_ecs',
-      method: 'amazonLinux2',
+  const myAsg = typedTemplate.resource('MyFleet');
+  expect(myAsg).toEqual(
+    matchConstruct({
+      machineImage: expect.objectContaining({
+        type: 'staticMethodCall',
+        fqn: 'aws-cdk-lib.aws_ecs.EcsOptimizedImage',
+        namespace: 'aws_ecs',
+        method: 'amazonLinux2',
+      }),
     })
   );
   expect(template.template).toBeValidTemplate();
