+++ test/type-resolution/enums.test.ts
@@ -1,9 +1,7 @@
 import * as reflect from 'jsii-reflect';
 import { Template } from '../../src/parser/template';
-import { StructExpression } from '../../src/type-resolution/struct';
 import { TypedTemplate } from '../../src/type-resolution/template';
-import { getCdkConstruct } from '../template';
-import { Testing } from '../util';
+import { matchConstruct, Testing } from '../util';
 
 let typeSystem: reflect.TypeSystem;
 
@@ -27,12 +25,11 @@ test('Enums are resolved correctly', async () => {
   const typedTemplate = new TypedTemplate(template, { typeSystem });
 
   // THEN
-  expect(template.template).toBeValidTemplate();
-  const myQueue = getCdkConstruct(typedTemplate, 'MyQueue');
-  expect(myQueue.type).toBe('construct');
-  expect(myQueue.props?.type).toBe('struct');
-  expect((myQueue.props as StructExpression)?.fields).toHaveProperty(
-    'encryption',
-    expect.objectContaining({ type: 'enum', choice: 'KMS' })
+  const myQueue = typedTemplate.resource('MyQueue');
+  expect(myQueue).toEqual(
+    matchConstruct({
+      encryption: expect.objectContaining({ type: 'enum', choice: 'KMS' }),
+    })
   );
+  expect(template.template).toBeValidTemplate();
 });
