+++ test/type-resolution/references.test.ts
@@ -1,9 +1,7 @@
 import * as reflect from 'jsii-reflect';
 import { Template } from '../../src/parser/template';
-import { StructExpression } from '../../src/type-resolution/struct';
 import { TypedTemplate } from '../../src/type-resolution/template';
-import { getCdkConstruct } from '../template';
-import { Testing } from '../util';
+import { matchConstruct, Testing } from '../util';
 
 let typeSystem: reflect.TypeSystem;
 
@@ -46,15 +44,14 @@ test('Constructs can be referenced', async () => {
   const typedTemplate = new TypedTemplate(template, { typeSystem });
 
   // THEN
-  const myApi = getCdkConstruct(typedTemplate, 'MyApi');
-  expect(myApi.type).toBe('construct');
-  expect(myApi.props?.type).toBe('struct');
-  expect((myApi.props as StructExpression)?.fields).toHaveProperty(
-    'handler',
-    expect.objectContaining({
-      type: 'intrinsic',
-      fn: 'ref',
-      logicalId: 'MyLambda',
+  const myApi = typedTemplate.resource('MyApi');
+  expect(myApi).toEqual(
+    matchConstruct({
+      handler: expect.objectContaining({
+        type: 'intrinsic',
+        fn: 'ref',
+        logicalId: 'MyLambda',
+      }),
     })
   );
   expect(template.template).toBeValidTemplate();
